+++ rotator/rotator.py
@@ -5,15 +5,25 @@ import json
 import logging
 import os
 import requests
-import urllib.parse
+from datetime import datetime, timezone
+from uritemplate import URITemplate
+from urllib.parse import urljoin
 
 
 # Client Secrets may include "numbers, letters and _, -, +, =, . symbols"
-EXCLUDE_CHARACTERS = r'''!"#$%&'()*,/:;<>?@[\]^`{|}~'''
+# …but the Client Registry has additional restrictions.
+EXCLUDE_CHARACTERS = r'''"%!'()*,/:;?@[\]`{|}~<>^&#$'''
+
+CLIENT_REGISTRY_TEMPLATE = URITemplate('https://auth0.cimpress.io/v1/clients/{client_id}/secrets')
+
+REQUIRED_FIELDS = ['id', 'secret']
 
 
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
+
+# Setup the client
+service_client = boto3.client('secretsmanager', endpoint_url=os.environ['SECRETS_MANAGER_ENDPOINT'])
 
 
 def lambda_handler(event, context):
@@ -34,42 +44,36 @@ def lambda_handler(event, context):
     token = event['ClientRequestToken']
     step = event['Step']
 
-    # Setup the client
-    service_client = boto3.client('secretsmanager', endpoint_url=os.environ['SECRETS_MANAGER_ENDPOINT'])
-
     # Make sure the version is staged correctly
     metadata = service_client.describe_secret(SecretId=arn)
     if not metadata['RotationEnabled']:
-        logger.error("Secret %s is not enabled for rotation" % arn)
-        raise ValueError("Secret %s is not enabled for rotation" % arn)
+        logger.error(f'Secret {arn} is not enabled for rotation')
+        raise ValueError(f'Secret {arn} is not enabled for rotation')
     versions = metadata['VersionIdsToStages']
     if token not in versions:
-        logger.error("Secret version %s has no stage for rotation of secret %s." % (token, arn))
-        raise ValueError("Secret version %s has no stage for rotation of secret %s." % (token, arn))
-    if "AWSCURRENT" in versions[token]:
-        logger.info("Secret version %s already set as AWSCURRENT for secret %s." % (token, arn))
+        logger.error(f'Secret version {token} has no stage for rotation of secret {arn}.')
+        raise ValueError(f'Secret version {token} has no stage for rotation of secret {arn}.')
+
+    if 'AWSCURRENT' in versions[token]:
+        logger.info(f'Secret version {token} already set as AWSCURRENT for secret {arn}.')
         return
-    elif "AWSPENDING" not in versions[token]:
-        logger.error("Secret version %s not set as AWSPENDING for rotation of secret %s." % (token, arn))
-        raise ValueError("Secret version %s not set as AWSPENDING for rotation of secret %s." % (token, arn))
+    elif 'AWSPENDING' not in versions[token]:
+        logger.error(f'Secret version {token} not set as AWSPENDING for rotation of secret {arn}.')
+        raise ValueError(f'Secret version {token} not set as AWSPENDING for rotation of secret {arn}.')
 
-    if step == "createSecret":
-        create_secret(service_client, arn, token)
-
-    elif step == "setSecret":
-        set_secret(service_client, arn, token)
-
-    elif step == "testSecret":
-        test_secret(service_client, arn, token)
-
-    elif step == "finishSecret":
-        finish_secret(service_client, arn, token)
-
+    if step == 'createSecret':
+        create_secret(arn, token)
+    elif step == 'setSecret':
+        set_secret(arn, token)
+    elif step == 'testSecret':
+        test_secret(arn, token)
+    elif step == 'finishSecret':
+        finish_secret(arn, token)
     else:
-        raise ValueError("Invalid step parameter")
+        raise ValueError('Invalid step parameter')
 
 
-def create_secret(service_client, arn, token):
+def create_secret(arn, token):
     """Create the secret
     This method first checks for the existence of a secret for the passed in token. If one does not exist, it will generate a
     new secret and put it with the passed in token.
@@ -81,13 +85,12 @@ def create_secret(service_client, arn, token):
         ResourceNotFoundException: If the secret with the specified arn and stage does not exist
     """
     # Make sure the current secret exists
-    current_dict = _get_secret_dict(service_client, arn, "AWSCURRENT")
+    current_dict = _get_secret_dict(arn, 'AWSCURRENT')
 
     # Now try to get the secret version, if that fails, put a new secret
     try:
-        _get_secret_dict(service_client, arn, "AWSPENDING", token)
-        logger.info("createSecret: Successfully retrieved secret for %s." % arn)
-
+        _get_secret_dict(arn, 'AWSPENDING', token)
+        logger.info(f'createSecret: Successfully retrieved secret for {arn}.')
     except service_client.exceptions.ResourceNotFoundException:
         # Generate a random password
         client_secret = service_client.get_random_password(PasswordLength=64, ExcludeCharacters=EXCLUDE_CHARACTERS)
@@ -95,10 +98,10 @@ def create_secret(service_client, arn, token):
 
         # Put the secret
         service_client.put_secret_value(SecretId=arn, ClientRequestToken=token, SecretString=json.dumps(current_dict), VersionStages=['AWSPENDING'])
-        logger.info("createSecret: Successfully put secret for ARN %s and version %s." % (arn, token))
+        logger.info(f'createSecret: Successfully put secret for ARN {arn} and version {token}.')
 
 
-def set_secret(service_client, arn, token):
+def set_secret(arn, token):
     """Set the secret
     This method should set the AWSPENDING secret in the service that the secret belongs to. For example, if the secret is a database
     credential, this method should take the value of the AWSPENDING secret and set the user's password to this value in the database.
@@ -108,19 +111,18 @@ def set_secret(service_client, arn, token):
         token (string): The ClientRequestToken associated with the secret version
     """
     # First try to create an access token with the pending secret. If it succeeds, return
-    pending_dict = _get_secret_dict(service_client, arn, 'AWSPENDING', token)
+    pending_dict = _get_secret_dict(arn, 'AWSPENDING', token)
     access_token = _create_access_token(pending_dict)
     if access_token:
         logger.info(f'set_secret: AWSPENDING secret is already set as client secret for secret {arn}.')
         return
 
     # Now try the current secret
-    access_token = _create_access_token(_get_secret_dict(service_client, arn, 'AWSCURRENT'))
+    access_token = _create_access_token(_get_secret_dict(arn, 'AWSCURRENT'))
     if not access_token:
     # If both current and pending do not work, try previous
         try:
-            access_token = _create_access_token(_get_secret_dict(service_client, arn, 'AWSPREVIOUS'))
-
+            access_token = _create_access_token(_get_secret_dict(arn, 'AWSPREVIOUS'))
         except service_client.exceptions.ResourceNotFoundException:
             access_token = None
 
@@ -132,7 +134,7 @@ def set_secret(service_client, arn, token):
     _set_client_secret(pending_dict, access_token)
 
 
-def test_secret(service_client, arn, token):
+def test_secret(arn, token):
     """Test the secret
     This method should validate that the AWSPENDING secret works in the service that the secret belongs to. For example, if the secret
     is a database credential, this method should validate that the user can login with the password in AWSPENDING and that the user has
@@ -143,12 +145,12 @@ def test_secret(service_client, arn, token):
         token (string): The ClientRequestToken associated with the secret version
     """
     # Try to acquire an acccess token with the pending secret
-    access_token = _create_access_token(_get_secret_dict(service_client, arn, 'AWSPENDING', token))
+    access_token = _create_access_token(_get_secret_dict(arn, 'AWSPENDING', token))
     if not access_token:
         raise ValueError(f'test_secret: Unable to acquire access token with pending secret of secret ARN {arn}.')
 
 
-def finish_secret(service_client, arn, token):
+def finish_secret(arn, token):
     """Finish the secret
     This method finalizes the rotation process by marking the secret version passed in as the AWSCURRENT secret.
     Args:
@@ -161,18 +163,18 @@ def finish_secret(service_client, arn, token):
     # First describe the secret to get the current version
     metadata = service_client.describe_secret(SecretId=arn)
     current_version = None
-    for version in metadata["VersionIdsToStages"]:
-        if "AWSCURRENT" in metadata["VersionIdsToStages"][version]:
+    for version in metadata['VersionIdsToStages']:
+        if 'AWSCURRENT' in metadata['VersionIdsToStages'][version]:
             if version == token:
                 # The correct version is already marked as current, return
-                logger.info("finishSecret: Version %s already marked as AWSCURRENT for %s" % (version, arn))
+                logger.info(f'finishSecret: Version {version} already marked as AWSCURRENT for {arn}')
                 return
             current_version = version
             break
 
     # Finalize by staging the secret version current
-    service_client.update_secret_version_stage(SecretId=arn, VersionStage="AWSCURRENT", MoveToVersionId=token, RemoveFromVersionId=current_version)
-    logger.info("finishSecret: Successfully set AWSCURRENT stage to version %s for secret %s." % (token, arn))
+    service_client.update_secret_version_stage(SecretId=arn, VersionStage='AWSCURRENT', MoveToVersionId=token, RemoveFromVersionId=current_version)
+    logger.info(f'finishSecret: Successfully set AWSCURRENT stage to version {token} for secret {arn}.')
 
 
 def _create_access_token(secret_dict):
@@ -210,21 +212,21 @@ def _create_access_token(secret_dict):
 
 
 def _get_openid_configuration():
-    """Gets the OpenID configuration for an authority
+    """Gets the OpenID configuration for an issuer
 
     This helper function retrieves the OpenID configuration from the well-known address.
     This configuration includes the endpoint at which client_credentials flows can
     be performed.
 
     Returns:
-        Configuration: The OpenID configuration for the configured authority.
+        Configuration: The OpenID configuration for the configured issuer.
 
     """
     headers = {
         'Accept': 'application/json'
     }
 
-    url = urllib.parse.urljoin(os.environ['AUTHORITY'], '/.well-known/openid-configuration')
+    url = urljoin(os.environ['ISSUER'], '/.well-known/openid-configuration')
     return requests.get(url, headers=headers).json()
 
 
@@ -234,7 +236,8 @@ def _set_client_secret(secret_dict, access_token):
     This helper function sets the client secret
     """
     payload = {
-        'client_secret': secret_dict['secret']
+        'client_secret': secret_dict['secret'],
+        'expire_previous_secrets_at': datetime.now(timezone.utc),
     }
     headers = {
         'Accept': 'application/json',
@@ -242,15 +245,12 @@ def _set_client_secret(secret_dict, access_token):
         'Content-Type': 'application/json'
     }
 
-    # todo(cosborn)
-    # It would be ideal to use the Auth0 Management API, but they lack robust permissions --
-    # I'd be able to rotate anybody's secret.
-    url = urllib.parse.urljoin('https://auth0.cimpress.io/v1/clients/', secret_dict['id'])
-    response = requests.patch(url, headers=headers, json=payload)
+    url = CLIENT_REGISTRY_TEMPLATE.expand(client_id=secret_dict['id'])
+    response = requests.post(url, headers=headers, json=payload)
     response.raise_for_status()
 
 
-def _get_secret_dict(service_client, arn, stage, token=None):
+def _get_secret_dict(arn, stage, token=None):
     """Gets the secret dictionary corresponding to the secret arn, stage, and token
 
     This helper function gets client credentials for the arn and stage passed in and returns the dictionary by parsing the JSON string
@@ -269,7 +269,6 @@ def _get_secret_dict(service_client, arn, stage, token=None):
         ValueError: If the secret is not valid JSON
 
     """
-    required_fields = ['id', 'secret']
 
     # Only do VersionId validation against the stage if a token is passed in
     if token:
@@ -280,7 +279,7 @@ def _get_secret_dict(service_client, arn, stage, token=None):
     secret_dict = json.loads(plaintext)
 
     # Run validations against the secret
-    for field in required_fields:
+    for field in REQUIRED_FIELDS:
         if field not in secret_dict:
             raise KeyError(f'{field} key is missing from secret JSON')